+++ src/main/java/sample/aircraft_type/AircraftTypeDAO.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package sample.aircraftType;
+package sample.aircraft_type;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
