+++ src/main/java/sample/controllers/BookingController.java
@@ -18,14 +18,14 @@ import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 import sample.aircraft.AircraftDAO;
 import sample.aircraft.AircraftDTO;
-import sample.aircraftType.AircraftTypeDAO;
-import sample.aircraftType.AircraftTypeDTO;
+import sample.aircraft_type.AircraftTypeDAO;
+import sample.aircraft_type.AircraftTypeDTO;
 import sample.airport.AirportDAO;
 import sample.airport.AirportDTO;
 import sample.booking.BookingDAO;
 import sample.booking.BookingDTO;
-import sample.bookingDetail.BookingDetailDAO;
-import sample.bookingDetail.BookingDetailDTO;
+import sample.booking_detail.BookingDetailDAO;
+import sample.booking_detail.BookingDetailDTO;
 import sample.flight.FlightDAO;
 import sample.flight.FlightDTO;
 import sample.members.MemberDAO;
