+++ src/main/java/sample/controllers/FlightController.java
@@ -22,8 +22,8 @@ import sample.aircraft.AircraftDTO;
 import sample.airport.AirportDAO;
 import sample.airport.AirportDTO;
 import sample.booking.BookingDAO;
-import sample.bookingDetail.BookingDetailDAO;
-import sample.bookingDetail.BookingDetailDTO;
+import sample.booking_detail.BookingDetailDAO;
+import sample.booking_detail.BookingDetailDTO;
 import sample.flight.FlightDAO;
 import sample.flight.FlightDTO;
 import sample.passengers.PassengerDAO;
